Refactor surface contracts out of validate modules
