+++ graphistry/PlotterBase.py
@@ -4,7 +4,8 @@ from typing_extensions import Literal
 from graphistry.io.types import ComplexEncodingsDict
 from graphistry.models.collections import CollectionsInput
 from graphistry.models.types import ValidationMode, ValidationParam
-from graphistry.validate import URLParamsDict, ReactSettingsDict
+from graphistry.models.surfaces.graphistry_react import ReactSettingsDict
+from graphistry.models.surfaces.graphistry_url import URLParamsDict
 from graphistry.validate.validate_react_encodings import parse_apply_encodings_ops
 from graphistry.plugins_types.hypergraph import HypergraphResult
 from graphistry.render.resolve_render_mode import resolve_render_mode
