+++ graphistry/models/surfaces/__init__.py
@@ -1,0 +1,36 @@
+"""Typed contracts for Graphistry-facing surfaces."""
+
+from .graphistry_axis import (  # noqa: F401
+    AXIS_BOUNDS_ALLOWED_KEYS,
+    AXIS_ROW_ALLOWED_KEYS,
+    AXIS_ROW_BOOL_KEYS,
+    AXIS_ROW_NUMERIC_KEYS,
+    AXIS_ROW_POSITION_KEYS,
+    AxisBounds,
+    AxisKind,
+    AxisRow,
+    AxisRows,
+    RingCategoricalAxis,
+    RingContinuousAxis,
+)
+from .graphistry_react import (  # noqa: F401
+    APPLY_ENCODINGS_REACT_KEYS,
+    APPLY_ENCODINGS_REACT_KEY_SET,
+    KnownReactSettingsDict,
+    REACT_SETTING_NAMES,
+    REACT_SETTING_NAME_SET,
+    ReactSettingsDict,
+    apply_encodings_keys,
+    react_setting_keys,
+)
+from .graphistry_url import (  # noqa: F401
+    LINEAR_AXIS_URL_DEFAULTS,
+    RADIAL_AXIS_URL_DEFAULTS,
+    URL_PARAM_NAMES,
+    URL_PARAM_NAME_SET,
+    URLParamsDict,
+    KnownURLParamsDict,
+    axis_url_defaults,
+    url_param_keys,
+)
+from .settings_value import SettingsValue  # noqa: F401
